move retry_after() to lib
